requirements.txt
pyPDF2==3.0.1
python-dotenv==1.0.0
streamlit
faiss-cpu==1.7.4
streamlit-extras
langchain[all]
transformers
pandas
pdf2image
pillow
numpy
pytesseract
opencv-python-headless==4.5.5.64